controller de la view inicial único
