+++ app/Logo.php
@@ -6,5 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class Logo extends Model
 {
-    //
+    protected $fillable = [
+        'logo', 
+    ];
 }
