+++ public/css/styleshirt.css
@@ -64,7 +64,7 @@ section.pricing {
     margin-right:1%;
     width:29px;
     height:22px;
-    
+   
   }
